% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignHhVehicleDvmtSplit.R
\docType{data}
\name{AssignHhVehicleDvmtSplitSpecifications}
\alias{AssignHhVehicleDvmtSplitSpecifications}
\title{Specifications list for AssignHhVehicleDvmtSplit module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignHhVehicleDvmtSplit.R script.
}
\usage{
AssignHhVehicleDvmtSplitSpecifications
}
\description{
A list containing specifications for the AssignHhVehicleDvmtSplit module.
}
\keyword{datasets}
