% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignHhVehiclePowertrain.R
\name{AssignHhVehiclePowertrain}
\alias{AssignHhVehiclePowertrain}
\title{Assign powertrain types to household vehicles.}
\usage{
AssignHhVehiclePowertrain(L, M)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}

\item{M}{A list the module functions of modules called by this module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignHhVehiclePowertrain} assigns the powertrain type to each household
vehicle.
}
\details{
This function assigns the powertrain type (ICEV, HEV, PHEV, BEV) to each
household vehicle.
}
