% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignHhVehiclePowertrain.R
\docType{data}
\name{AssignHhVehiclePowertrainSpecifications}
\alias{AssignHhVehiclePowertrainSpecifications}
\title{Specifications list for AssignHhVehiclePowertrain module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{model inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignHhVehiclePowertrain.R script.
}
\usage{
AssignHhVehiclePowertrainSpecifications
}
\description{
A list containing specifications for the AssignHhVehiclePowertrain module.
}
\keyword{datasets}
