% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCarbonIntensity.R
\name{calcAverageFuelCI}
\alias{calcAverageFuelCI}
\title{Calculate average fuel carbon intensity of a transportation mode and type}
\usage{
calcAverageFuelCI(FuelCI_, FuelProp_, BiofuelProp_)
}
\arguments{
\item{FuelCI_}{a named numeric vector of carbon intensity of fuel types where
the values are grams of carbon dioxide equivalents per megajoule and the
names are Gasoline, Diesel, Cng (compressed natural gas), Lng (liquified
natural gas), Ethanol, Biodiesel, and Rng (renewable natural gas).}

\item{FuelProp_}{a named vector of fuel proportions used by the mode and
type, or in the case of transit with multiple metropolitan area data, a
matrix of fuel proportions by type and metropolitan area. The names must be
the names of the base fuel types consistent with the names used in FuelCI_
although only the names of fuels used by the mode and type need to be
included.}

\item{BiofuelProp_}{a named vector of the biofuel proportions of base fuels,
or in the case of transit with multiple metropolitan area data, a matrix
of biofuel proportions by type and metropolitan area. The names must be in
form of the biofuel name concatenated with 'Prop' and concatenated with the
base fuel name (e.g. EthanolPropGasoline).}
}
\description{
\code{calcAverageFuelCI} calculates the average carbon intensity of fuels
used by a transportation mode and type considering the carbon intensities of
the base fuels, biofuel mixtures, and the proportions of fuels used.
}
\details{
The function calculates the average carbon intensity of fuels used by a
transportation mode (e.g. household, car service, commercial service, public
transit, freight) and type (e.g. auto, light truck, van, bus, rail, heavy
truck). The average carbon intensity is calculated from the base fuel mix
for the mode and type (e.g. gasoline, diesel, compressed natural gas), the
mix of biofuels used for the mode and type (e.g. ethanol mix in gasoline),
and the mix of powertrains geared to the different base fuel types (e.g.
proportion of light-duty vehicles that run on gasoline vs. the proportion
running on diesel).
}
