% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateComEnergyAndEmissions.R
\docType{data}
\name{CalculateComEnergyAndEmissionsSpecifications}
\alias{CalculateComEnergyAndEmissionsSpecifications}
\title{Specifications list for CalculateComEnergyAndEmissions module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateComEnergyAndEmissions.R script.
}
\usage{
CalculateComEnergyAndEmissionsSpecifications
}
\description{
A list containing specifications for the CalculateComEnergyAndEmissions
module.
}
\keyword{datasets}
