% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHhEnergyAndEmissions.R
\name{CalculateHhEnergyAndEmissions}
\alias{CalculateHhEnergyAndEmissions}
\title{Calculate energy and emissions of household vehicle travel.}
\usage{
CalculateHhEnergyAndEmissions(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateHhEnergyAndEmissions} calculates the hydrocarbon and
electrical energy consumption of the travel of each household vehicle and the
carbon emissions produced.
}
\details{
This function calculates the hydrocarbon and electrical energy consumption of
the travel of each household vehicle and the carbon emissions produced. It
assigns households to be eco-driving households or not if that hasn't been
done before.
}
