% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHhEnergyAndEmissions.R
\docType{data}
\name{CalculateHhEnergyAndEmissionsSpecifications}
\alias{CalculateHhEnergyAndEmissionsSpecifications}
\title{Specifications list for CalculateHouseholdEnergyAndEmissions module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateHhEnergyAndEmissions.R script.
}
\usage{
CalculateHhEnergyAndEmissionsSpecifications
}
\description{
A list containing specifications for the CalculateHouseholdEnergyAndEmissions
module.
}
\keyword{datasets}
