% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMpgMpkwhAdjustments.R
\name{CalculateMpgMpkwhAdjustments}
\alias{CalculateMpgMpkwhAdjustments}
\title{Main function to calculate MPG/MPkWh adjustment factors.}
\usage{
CalculateMpgMpkwhAdjustments(L)
}
\arguments{
\item{L}{A list containing data requested by the module from the datastore.}
}
\value{
A list containing data identified in the module Set specifications.
}
\description{
\code{CalculateMpgMpkwhAdjustments} calculates MPG and MPkWh adjustment
factors for the effects of speed smoothing, eco-driving, and congestion.
}
\details{
This function calculates factors for adjusting the fuel economy (MPG) and
electric energy efficiency (MPkWh) of vehicles of different type in different
Mareas as due to speed smoothing (i.e. reducing speed variation as a result
of active traffic management measures), eco-driving behavior, and roadway
congestion.
}
