% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCarbonIntensity.R
\name{filterOnNames}
\alias{filterOnNames}
\title{Filter named, vector, list, data frame, or matrix based on names containing
key 'words'}
\usage{
filterOnNames(Data_misc, Filter_, Remove_ = NULL)
}
\arguments{
\item{Data_misc}{a named vector, list, data frame, or matrix}

\item{Filter_}{a string vector of 'words' to find in the names}

\item{Remove_}{a string vector of 'words' to remove from the names after the
filtering has been completed.}
}
\value{
An object of same class as the input with the only the elements or
columns that contain all the key 'words'
}
\description{
\code{filterOnNames} selects portion of a vector, list, data frame, or
matrix whose names (colnames in case of matrix) contain the all the 'words'
in the supplied vector of 'words'.
}
\details{
This function filters a named vector, list, data frame or matrix based on
whether a vector of key 'words' are in all the names.
}
