% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignHhVehicleDvmtSplit.R
\docType{data}
\name{VmtSplitModel}
\alias{VmtSplitModel}
\title{Household vehicle DVMT split model}
\format{A string:
\describe{
  \item{VmtSplitModel}{a string representation of the estimated linear model
  which splits DVMT among household vehicles}
}}
\source{
AssignHhVehicleDvmtSplit.R script.
}
\usage{
VmtSplitModel
}
\description{
A string representation of the estimated linear model which splits DVMT among
household vehicles.
}
\keyword{datasets}
